unit tests, added authorization, api policies

+++ app/Http/Controllers/ContactsController.php
@@ -8,23 +8,28 @@ use App\Contact;
 class ContactsController extends Controller
 {
     //
+    public function index() {
+        $this->authorize('viewAny', Contact::class);
+        return request()->user()->contacts;
+    }
+
     public function store() {
-
-
-        Contact::create($this->validateData());
+        $this->authorize('create', Contact::class);
+        request()->user()->contacts()->create($this->validateData());
     }
 
     public function show(Contact $contact) {
+        $this->authorize('view', $contact);
         return $contact;
     }
 
     public function update(Contact $contact) {
-
+        $this->authorize('update', $contact);
         $contact->update($this->validateData());
     }
 
     public function destroy(Contact $contact) {
-
+        $this->authorize('delete', $contact);
         $contact->delete();
     }
